Refactor internal ThreadPool.h into halide_thread_pool.h tool
