+++ src/CMakeLists.txt
@@ -152,7 +152,6 @@ set(HEADER_FILES
     StrictifyFloat.h
     Substitute.h
     Target.h
-    ThreadPool.h
     Tracing.h
     TrimNoOps.h
     Tuple.h
