+++ src/autoschedulers/adams2019/AutoSchedule.cpp
@@ -72,6 +72,8 @@
 #ifdef _WIN32
 #include <io.h>
 #define _isatty isatty;
+#else
+#include <unistd.h>
 #endif
 
 namespace Halide {
