+++ tools/CMakeLists.txt
@@ -48,3 +48,9 @@ add_library(Halide_Tools INTERFACE)
 add_library(Halide::Tools ALIAS Halide_Tools)
 target_include_directories(Halide_Tools INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
 set_target_properties(Halide_Tools PROPERTIES EXPORT_NAME Tools)
+
+add_library(Halide_ThreadPool INTERFACE)
+add_library(Halide::ThreadPool ALIAS Halide_ThreadPool)
+target_link_libraries(Halide_ThreadPool INTERFACE Threads::Threads)
+target_include_directories(Halide_Tools INTERFACE $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}>)
+set_target_properties(Halide_ThreadPool PROPERTIES EXPORT_NAME ThreadPool)
